Move merge_anatomical_morphological_final_for_pred function to read_files.py
